fix: txn table
