+++ apps/storybook/src/stories/transactions/TransactionTable.stories.ts
@@ -1,10 +1,4 @@
 import { TransactionTable } from "@geist/ui-react/components/transactions/transaction-table";
-import { asTransactionMeta } from "@geist/ui-react/lib/blockscout/api";
-import {
-	TXN_LIST,
-	generateTxnFixturesByCount,
-} from "@geist/ui-react/lib/blockscout/data.fixture";
-import { Explorer } from "@geist/ui-react/lib/explorer/url";
 import type { Meta, StoryObj } from "@storybook/react";
 import { withWagmiProvider } from "../decorators/wagmi";
 
@@ -15,45 +9,24 @@ const meta = {
 	parameters: {
 		layout: "centered",
 	},
-	decorators: [withWagmiProvider()],
-	argTypes: {
-		explorer: { options: Object.values(Explorer) },
-	},
 } satisfies Meta<typeof TransactionTable>;
 
 export default meta;
+
 type Story = StoryObj<typeof meta>;
 
-export const TransactionTableSome: Story = {
+export const ContractCreationAndContractCall: Story = {
 	args: {
-		transactions: TXN_LIST.map((txn) => asTransactionMeta(txn)),
-	},
-};
-
-export const TransactionTableMany: Story = {
-	args: {
-		transactions: generateTxnFixturesByCount(50).map((txn) =>
-			asTransactionMeta(txn),
-		),
-	},
-};
-
-export const TransactionTableWithBlockscoutExplorerLink: Story = {
-	args: {
-		transactions: generateTxnFixturesByCount(50).map((txn) =>
-			asTransactionMeta(txn),
-		),
-		explorer: Explorer.Blockscout,
+		type: ["contract_creation", "contract_call"],
 		chainId: 1,
 	},
+	decorators: [withWagmiProvider()],
 };
 
-export const TransactionTableWithBlockscoutExplorerLinkInOptimism: Story = {
+export const ContractCreationAndContractCallInOPMainnet: Story = {
 	args: {
-		transactions: generateTxnFixturesByCount(50).map((txn) =>
-			asTransactionMeta(txn),
-		),
-		explorer: Explorer.Blockscout,
+		type: ["contract_creation", "contract_call"],
 		chainId: 10,
 	},
+	decorators: [withWagmiProvider()],
 };
