+++ apps/storybook/src/stories/transactions/TransactionTableWithDetails.stories.ts
@@ -1,4 +1,10 @@
 import { TransactionTableWithDetails } from "@geist/ui-react/components/transactions/transaction-table-with-details";
+import { asTransactionMeta } from "@geist/ui-react/lib/blockscout/api";
+import {
+	TXN_LIST,
+	generateTxnFixturesByCount,
+} from "@geist/ui-react/lib/blockscout/data.fixture";
+import { Explorer } from "@geist/ui-react/lib/explorer/url";
 import type { Meta, StoryObj } from "@storybook/react";
 import { withWagmiProvider } from "../decorators/wagmi";
 
@@ -9,26 +15,45 @@ const meta = {
 	parameters: {
 		layout: "centered",
 	},
+	decorators: [withWagmiProvider()],
+	argTypes: {
+		explorer: { options: Object.values(Explorer) },
+	},
 } satisfies Meta<typeof TransactionTableWithDetails>;
 
 export default meta;
-
 type Story = StoryObj<typeof meta>;
 
-export const TransactionTableWithDetailsContractCreationAndContractCall: Story =
-	{
-		args: {
-			type: ["contract_creation", "contract_call"],
-			chainId: 1,
-		},
-		decorators: [withWagmiProvider()],
-	};
+export const TransactionTableSome: Story = {
+	args: {
+		transactions: TXN_LIST.map((txn) => asTransactionMeta(txn)),
+	},
+};
 
-export const TransactionTableWithDetailsContractCreationAndContractCallInOPMainnet: Story =
-	{
-		args: {
-			type: ["contract_creation", "contract_call"],
-			chainId: 10,
-		},
-		decorators: [withWagmiProvider()],
-	};
+export const TransactionTableMany: Story = {
+	args: {
+		transactions: generateTxnFixturesByCount(50).map((txn) =>
+			asTransactionMeta(txn),
+		),
+	},
+};
+
+export const TransactionTableWithBlockscoutExplorerLink: Story = {
+	args: {
+		transactions: generateTxnFixturesByCount(50).map((txn) =>
+			asTransactionMeta(txn),
+		),
+		explorer: Explorer.Blockscout,
+		chainId: 1,
+	},
+};
+
+export const TransactionTableWithBlockscoutExplorerLinkInOptimism: Story = {
+	args: {
+		transactions: generateTxnFixturesByCount(50).map((txn) =>
+			asTransactionMeta(txn),
+		),
+		explorer: Explorer.Blockscout,
+		chainId: 10,
+	},
+};
